moving `contact.helper.js` to features/persons subdir
